+++ lib/MobileRTC.framework/Headers/MobileRTCMeetingSettings.h
@@ -81,12 +81,6 @@
 @property (assign, nonatomic) BOOL disconnectAudioHidden;
 
 /*!
- @brief Enable/Disable Kubi Device in the meeting.
- @warning The option is available only on iPad if you want to use Kubi device. 
- */
-@property (assign, nonatomic) BOOL enableKubi;
-
-/*!
  @brief Change thumbnail video layout while viewing a share in the meeting.
  @warning If you set it to YES, the video of attendees will be placed at right of the Landscape(the device screen is oriented horizontally) or the bottom of Portrait(the device screen is oriented vertically) apart from the shared content, which means the video won't cover the content; if you set to NO, it will show only the video of active speaker and the video will be placed in the bottom right of the screen.
  */
@@ -272,6 +266,12 @@
  @param enable YES means show meeting elapse time, otherwise not.
  */
 - (void)enableShowMyMeetingElapseTime:(BOOL)enable;
+
+/*!
+@brief Set the visibility of reaction on meeting UI. Default is displaying.
+@param hidden YES means hide reaction emotion.
+*/
+- (void)hideReactionsOnMeetingUI:(BOOL)hidden;
 
 /*!
 @brief pre populate webinar registration info.